ONC-11999 Move source files to a new package directory
