+++ src/llmvalidate/utils.py
@@ -1,6 +1,6 @@
 from typing import Any, List, Dict, Optional, Union
 from ast import literal_eval
-from src.structured import StructuredResult
+from src.llmvalidate.structured import StructuredResult
 import pandas as pd
 import re
 import json
