+++ tests/compare_results_all_test.py
@@ -1,7 +1,7 @@
-﻿import math
+﻿﻿import math
 import pandas as pd
 import pytest
-import src.validation as v
+import src.llmvalidate.validation as v
 pd.options.display.width = 0
 
 def get_test_df(addconfidence):
