+++ tests/compare_results_test.py
@@ -1,5 +1,5 @@
 from numpy import nan
-import src.validation as v
+import src.llmvalidate.validation as v
 import pytest
 import math
