+++ tests/validate_test.py
@@ -1,9 +1,9 @@
-﻿import pandas as pd
+﻿﻿import pandas as pd
 import pytest
-import src.validation as v
-from src.utils import convert_lists
-from src.structured import StructuredResult, StructuredField, StructuredGroup 
-from src.utils import flatten_structured_result
+import src.llmvalidate.validation as v
+from src.llmvalidate.utils import convert_lists
+from src.llmvalidate.structured import StructuredResult, StructuredField, StructuredGroup 
+from src.llmvalidate.utils import flatten_structured_result
 pd.options.display.width = 0
 
 _CASES = [
